Warning fix

+++ src/AddressTree.cpp
@@ -467,7 +467,7 @@ void AddressTree::getAddresses(lua_State *vm) {
     snprintf(key, sizeof(key), "%02X:%02X:%02X:%02X:%02X:%02X", mac[0], mac[1],
              mac[2], mac[3], mac[4], mac[5]);
 
-    snprintf(val, sizeof(val), "%llu", it->second);
+    snprintf(val, sizeof(val), "%llu", (long long unsigned int)it->second);
 
     lua_push_str_table_entry(vm, key, val);
   }
